Tests for functionality in 9708921
It also includes three domain classes that are necessary when running
the tests in a memory database in order to generate the schema. These
domain classes will likely be important in the future, but for now
there's no commiting as to their contract (even their names). The plugin
otherwise works without them because the QueriesResourceService
generates and runs SQL queries that bypass GORM, as is required by the
i2b2 model.