Add QueryToolController
The QueryController deals with QueryDefinitions in XML instead of JSON
as could be expected. The reason is that in the i2b2 data model, the
query definition is stored as XML in the database. To avoid adding more
conversion routines, this time for JSON (we now have conversion routines
to and from XML), we just communicate with XML as well.