Fix QueriesResourceServiceTests on h2
Also flush before running SQL query; do not flush on the test. This
flushing is generally needed before running the SQL so we might as well
add it to the service.

With the addition of a relationship in QtPatientSetCollection to
PatientDimension, hibernate creates a foreign key. Therefore, we have to
add the patients to patient_dimension otherwise inserting into
qt_patient_set_collection will fail with a foreign key constraint
violation.