Fix behavior for RModules.transferImageFile=false
The case for when RModules.transferImageFile was broken in several ways:
for instance, it did not create the zip file or fill in the linksArray.

Unfortunately, I had to introduce a move step even when
transferImageFile is false. The reason is that the images will be under
<jobs_dir>/<job_name>/workingDirectory/. When transferImageFile is true,
they will be under <images dir>/<job_name>. The images url is aliased to
<jobs_dir> under the first case and <images dir> under the second case,
but the paths don't match. While initializeAttributes() gives back all
the image URLs through the linksArray parameter, this is ultimately
ignored by the code that displays acgh results, which expects the layout
used when transferImageFile is true. Consequently, I'm moving image
files under <job_name>/workingDirectory to its parent.