Remove dead code; simplify RModules config
Apparently some time ago you could have a different status list for
each analysis, and this information would go into the config file. But
this functionality was commented out. Remove it for good.
Additionally, since the status list is in fact static, define a static
variable with this list in the code. This reduces the amount of
useless configuration needed for Rmodules.