Fix write of wrong values in i2b2.c_operator
This only affectedt the i2b2metadata.i2b2 table, not the
i2b2.i2b2_secure table.
The c_operator table was having values 'N' and 'T' (indicating the
type of the data) being written to the c_oeprator table. But for
concept_dimensiopn ontology terms, this should always be 'LIKE'.