Fix arbitrary code execution in RModulesJobService
The only character being escaped was \. This is the best I could write
given my limited knowledge of R (I just tried following Rʼs lexer
code), should be reviewed.

This assumes that the R analyses are written so the params are inside
quotes (") or apostrophes ('). Otherwise, it's game over.

See THRONE-87.