Work around problem with H2 tests
For some reason, the tests stopped working. There's an exception when
the database connection is created saying the schemas being created in
the init file already exist. For reference, this is the data soure
definition:

	environments {
		test {
			dataSource {
				dbCreate = "update"
				url = "jdbc:h2:mem:testDb;MVCC=TRUE;LOCK_TIMEOUT=10000;INIT=" +
					"RUNSCRIPT FROM './h2_init.sql'"
			}
		}
	}

So I changed CREATE SCHEMA to CREATE SCHEMA IF NOT EXISTS.
I haven't gotten to the bottom of this (why it worked before, but not
now). Though I've noticed  the stack trace includes the creation of a
connection pool, so maybe there are several connections being made and
the second one finds the schemas already there when running h2_init.sql.