Update PatientSetQueryBuilderService.groovy
The facts containing the concept_cd 'SECURITY' should never be retreived (Patient set can contain patient from other nodes if not expluded). I do not know if this is the best place to put it, but I hope you will have a better idea. Another aproache would be to have some sort of EXCLUDE constraint which, than, would allow us to move the constraint into the JS file for subset query.