Ticket 227. Added LineGraph Y-axix labelling based on argument ‘HDD.data.type’
When argument has a value (default is NULL), it will be used as the y-axis label. Otherwise, the tail-end of the concept-path is used as a y-axis label (eg. ‘/Subject/Age/‘). In addition, the label will be suffixed by the type of statistic being plotted (ie. mean/median, se/sd).