Rename SingleOrMultiNumericVCC & change behavior
The change in behavior is that all the clinical data will always be
returned with a context (the values will be maps). In addition,
HighDimensionColumnConfigurator always includes the concept path in the
context in case multiConcepts is on. These two changes make the class
easier to use downstream by making the values that end up in BackingMap
more consistent in several cases.

However, the BoxPlot relying on having no context if there was only one
clinical variable, so it now has to switch between
ContextNumericVariableColumnConfigurator (the new name of the class) and
NumericColumnConfigurator.