Fix multi-highdim nodes for the same column
The previous code had a problem in that it looked for common *assays*
instead of common *patients*.

When implementing a fix it became clear that the strategy of combining
the TabularResults into one was too cumbersome. Consequently, it was
chanegd so that the HighDimensionMultipleRowsResultColumn now subscribes
multiple data sources. This resulted in simpler code.