Fix botched merge at d2c4316
The merge at d2c4316 was made with strategy ours (or equivalent) and
just discarded everything in the second parent, c1f358e, and kept the
code as in the first parent, 76b6ae4.

This commit attempts to correct this screw-up by properly merging
c1f358e into 76b6ae4 (and resolving the single conflict in
LineGraph.groovy), taking a diff against the botched merged and applying
the diff on top of the code as of the end of business on Friday,
f38d4ec.