Create view i2b2metadata.i2b2_trial_nodes
Establishes a map between study names and their top nodes.

This is needed for the efficient operation of the study-centric rest
API. Note that i2b2 would organize studies and their permissions in a
different manner: each study would have its own ontology table and a
specific entry in i2b2.table_access.

I added an index on i2b2metadata.i2b2 to speed up the view query in case
the table is very large. Perhaps a better option would be to add a
materialized view instead, but that is PostgreSQL 9.3 only and we're
still using 9.2 in quite a lot of places.