Fix travis build
The build files if when grails compile (or equivalent) is run, but
transmart-core-db is not in the Maven repository (classes of
transmart-core-db-test cannot find the core-db classes they depend on). I tried
to fix this by adding an explicit dependency of transmart-core-db-test on
core-db instead of just the inline plugin declaration. This had the odd effect
of having grails deleting the whole project directory when running grails
test-app on transmart-core-db-tests, so the that a dead end.

As a crappy workaround, I now separate thesteps of downloading the dependencies
and actually doing the compilation. So now the call to grails test-app is
preceded by a call to grails refresh-dependencies.