Fix extra data LineGraph
and possibly other analysis that used OptionalBinningColumnConfigurator
with keyForIsCategorical set. Problem was that setting the
MissingValueAction to a constant value to account for the case when no
variable is selected (see UAT-176) is too strong an action. If the
selected variable doesn't have data for some patients, those patients
would nevertheless show up (under SINGLE_GROUP).

Resolved this with an OptionalColumnConfiguratorDecorator, which is
anywat what I wanted to all along. This decorator delegates to two
different sub-configurators, selected based on whether some param key is
set.

Fixes UAT-177 (likely among other problems), previously broken by the
fix for UAT-176.