Adapt to security checks in transmartApp
transmartApp now needs to be able to get the "current user" at several
points. This is done by injecting the bean name "currentUserBean".
This bean is of type CurrentUserBeanFactoryBean, which returns a proxy
that tries to find several beans in order (with each one in a
different Spring scope). In RModules, we introduce a new scope, "job".
This scope is active inside Quartz threads whose entry point is
AnalysisQuartzJobAdapter::execute().

Because we 1) unfortunately, and perhaps unnecessarily, happen to call
some transmartApp services that end up using currentUserBean and 2) we
now implement an access check in Rmodules itself, we need to provide a
way to retrieve the current user (for 2) through the currentUserBean
mechanism (for 1).

This is done by, in the doWithApplicationContext hook, retrieving the
currentUserBean factory reference and registering a new bean to try.
The bean we are registering is job scoped.