Create secure object when loading private study
Management of study level access in the admin panel is actually the
management of what tranSMART calls "secure objects". These are stored in
the search_secure_object table. The permissions for these objects are
stored in search_auth_sec_object_access. At runtime, these two tables
are joined to check user access to a study (together with others, for
instance in order to handle user group membership). For that, the secure
object with type 'BIO_CLINICAL_TRIAL' and a token equal to the same that
is stored in i2b2_secure (typically `EXP:<trial name>`) is involved.

This commit adds the necessary row in search_secure_object. A row is
also inserted in bio_experiment for no other reason than the original
code does that. The bio_data_id of the secure object is then filled with
theh experiment id.

Original code:
https://github.com/transmart/transmart-data/blob/cbcdeb62a9aefe709fa11f75367fcdaeee9d07b6/ddl/postgres/tm_cz/functions/i2b2_secure_study.sql

Related: JE-49