Refactor external job configuration
Make it that instead of strict class hierarchy of parameters (for
instance, I see I'd need something like ExternalJobParameters <-
StudyBasedJobParameters <- StandardAsayMappingJobParameters <-
StandaardHighDimDataJobParameters <- MrnaJobParameters), we have
instead a final ExternalJobParameters that delegates to multiple
ExternalJobParametersModule:

    interaface ExternalJobParametersModule {

        Set<String> getSupportedParameters()

        void munge(ExternalJobParametersInternalInterface ejp)

        void validate(ExternalJobParametersInternalInterface ejp)
    }

Because we can delegate to several of these, we can mix and match the
parameters the way we want.

As a result of refactoring, the platform jobs now don't have the
STUDY_ID, TOP_NODE, and SECURITY_REQUIRED parameters like they used to
(because they were defined in the base ExternalJobParameters).