Fix role deletion error
Repro. steps:
Go to Role list, show role, and delete role
EXPECTED: After asking for confirmation, role should be removed from list
ACTUAL: Error occurs ("An error has occurred"),
role is not deleted. In addition, when clicking "detail"
on user of selected role, it tries to open the role instead of the user with the user's id, which fails