Stream observations response, use normalized vars
BC-breaking change!

Categorical variables are not flattened anymore. If a request for a
categorical variable concept comes in, the output will be observation
facts with type string directly under the categorical variable rather
than its leaves.

Additionally, we now don't load the full observations response in memory
and instead stream it.