Fix race condition in jobs
updateStatus and updateJobInputs are called from different threads at
almost the same point, updateJobInputs from the http thread and
updateStatus from the quartz thread (and later also from another http
thread as a response to checkstatus calls from the client, but this is
not relevant).

If the updateJobInputs transaction results is not visible on the
transaction on the quartz thread at the point where the quartz thread
issues the SELECT statement that precedes the UPDATE statement, then the
Hibernate session running on the quartz thread will cache stale data for
the input params column and will include this value in all the updates
thenceforth.

This fix is to move the update in the http thread (done in Rmodules) and
commit the transaction before the quartz thread code is run (done in
transmartApp).