@CompileStatic the database results
TabularResults and the associated iterators are now @CompileStatic

I made a small refactoring in using an AbstractIterator in CollectingTabularResult. There is a small behavior change in that previously the
first row would be loaded before returning the iterator and so errors in the first row would raise an exception immediately. Now the rows are
only loaded when calling hasNext() or next(). This behavior is not guaranteed in the api so it should not be a problem to change it.