Marge groovy and java source files (remove src/main/java dir)
According to upgrade guide: "It is recommended to merge Java source
files from src/java into src/main/groovy. You can create a src/main/java
directory if you want to and it will be used but it is generally better
to combine the folders. (The Groovy and Java sources compile together.)"