Remove @TupleConstructor from ProjectionMap
Groovy emits invalid bytecode for the constructor, but only sometimes. It started happening after converting IndexedArraySet from groovy to
java, which should not change this class in any way. TupleConstructor now wants to include the hidden metaClass field in the constructor which
it should not do. Maybe some kind of bug in @TupleConstructor or the order in which transformations run?