First implementation for modifiers (plan C: in-app join)
Initial implementation of modifier support in the hypercube. This implementation uses application side
joining to combine an observation with its modifiers. Initially this is a temporary implementation, the
goal is to push the joining to the database to improve efficiency and flexibility.

This also includes improvements to make ProjectionMap more robust.

Existing tests are working, but no tests for the new code yet, so probably not working!