Specify sexCd as string; refactor to i2b2data.PatientDimension from Patient
It's not necessary to create a custom PropertyImpl for sexCd.
SexCd is not specified in the Patient interface, and I'm not sure if it is a good idea to do so as I don't think other code should depend on
this property too much. So change the multidimquery.PatientDimension to use i2b2data.PatientDimension as element type (which is not exposed
from the rest-api, though the list of properties is).