Replace DimensionEqualator with IndexGetter
The IndexGetters are useful in more situations and can also be used to compare a value's dimensions.

Also add a determinesType property for Dimension's which is true for ConceptDimension and ProjectionDimension
which determine the type of the observation. This is needed to determine the sorting of a query automatically (not implemented).