Remove any remaining added copyright statements
Remove copyright statements in code that was created as part of the Transmart 17.1 project, as was agreed with the Transmart Foundation.
This comprises the newly written code in the history of commit 9d998de2d6c37f484e491733831ed97975c7e3e4 (which was tagged as the
transmart 17.1-deliverable), but excluding the Transmart 16.1/16.2 code in the separate git repositories which were merged into
this one (available in branch rebase-root, commit 58a48ff76a3b66396be28537c9ca6333b92cceb8, see README.md#Git history for details).

Copyright ownership of the code that was created as part of the Transmart 17.1 project is (as of this writing) unclear,
with The Hyve B.V. and the Transmart Foundation in disagreement on who of them owns this copyright. The Hyve has agreed with the
Transmart Foundation to remove the copyright notices in order to move the project to a close, with the actual copyright ownership
to be agreed upon at a later moment. However both parties agree to release the code under the GNU GPL version 3 or (at your option)
any later version. See the LICENSE file.