Clarify (lack of) support for null values
Null values, i.e. observations with both textValue and numberValue set to null, are now supported by the hypercube code and the protobuf
serialization. The database can store such values, but Transmart and the ETL tools do not support them. So the support in the serialization
is not used, it is purely for future compatibility.