Improve groupSamples, splitNullPackedDim, add tests
HypercubeProtobufSerializer.groupSamples should not assume that all pack elements are present in the group, but also
work if some of them are skipped. This change makes it scan through all element indexes.

This also changes moveNullPackedDimensionToFront into splitNullPackedDim which returns lists of null and non-null
pack element values instead of modifying the passed list. This works better with the rest of the code and should
be a bit more efficient.

This change does not quite fix the problems I ran into: if the very first group that we retrieve from a hypercube
has pack elements missing that are present in later groups, the order of element indexes will not correspond with
the sort order according to which the groups are ordered. I think the only solution is to not assume any order of
the pack elements. Still to be implemented.