groupSamples no longer depends on input ordering
It turns out we cannot rely on HypercubeValues to be ordered according to their packed dimension index. If the first row of values
does not include a value for a certain packed dim element X, then the hypercube won't be able to assign an index to X. Later rows
that do include values with element X will have their indexes for the packed dimension out of order as X will have a higher index.

Example:
first row contains
[val: ..., patient: patient_1], [val: ..., patient: patient_3], [val: ..., patient: patient_4]
The indexes will be assigned as patient_1: 0, patient_3: 1, patient_4: 2.
The next row, which is sorted according to the patient order by the database, contains
[val: ..., patient: patient_1], [val: ..., patient: patient_2], [val: ..., patient: patient_3], [val: ..., patient: patient_4]
patient_2 will be assigned index 3, and the second row no longer matches the order of the indexes.

There is no way to prevent this, as there can always be patients missing in early rows so the order in which the hypercube
sees the patients is not fixed. To fix the code, we make the HypercubeProtobufSerializer not depend on the order of patients and
we implement a kind of radix sort to get them in the right order.