Change return type of getPatientSetRequestConstraintsAndApiVersion to return tuple instead of list
Actually a purpose built @Immutable type.
Returning a list is easier in groovy, but is not type safe. For 'official' apis that try to be type safe such as core-api I prefer
doing a named type approach. Another option is to use grails.util.Pair, which is typesafe but doesn't give you descriptive names.

Also includes some whitespace changes and using '2.1-tests' as version in tests.