Maximize packed cell size
Add a HypercubeProtobufSerializer.MAXIMUM_PACKED_VALUES setting to limit the number of observations that get packed into a packed cell.
This is to prevent PackedCell protobuf messages from becoming too unwieldy. That could happen e.g. if values are retrieved with
only a single dimension (that is packable). Support for that is not yet implemented, but maximizing the size is good anyway.