SubSelectionConstraint for most dimensions
A generic SubSelectionConstraint (replacing the PatientSelectionConstraint) that works (in principle) for all dimensions.
At the moment only dimensions based on i2b2 dimensions are implemented, and only the patient dimension is tested. As this
functionality is needed quickly tests for other dimensions will follow.

Implements GBD-37, and implements without tests GDB-38 and GDB-39