Validate constraints immediately
Previously ConstraintFactory.create would throw a ConstraintBindingException for errors, but not for invalid values,
and it was up to the caller to check for validation problems. Now the exception is also thrown for validation problems
so the caller can't forget to check that.

Also improve some of the error messages.