Fix bug
IDEA believes that SubsetConstroller.log is an org.apache.commons.logging.Log, but at runtime it
is an slf4j logger, whose trace() method does not accept Object but String. That could be a
configuration problem or a deficiency in the IDEA grails plugin. Converting the json object to a
string solves the exception.