Save only the user given resultInstanceId description
Instead of having a description of "Patient set for \"${name}\"", we just use `name` as description.
The resultInstance.description does not appear to be used anywhere except in the rest api. TransmartApp
uses the queryMaster.name property which contains the same name but without the "Patient set for" wrapping.
This makes the query master name and the result instance description be identical.

Fixes GDB-71