Refactor DimensionImpl.selectDimensionElements
Have a generic implementation for the I2b2 dimensions, and a generic failing implementation for highdim dimemsions.

Counting for visits does not seem possible in the legacy hibernate criteria api, I can't find a way to count distinct on two properties.
The implementation for modifiers was not in accordance with the model of treating each modifier as a dimension, I took it out but didn't
put in a replacement yet.