Fix aggregate combining
Re-work the checks to see if multiple aggregates can be combined in a single db call. The code now is (I hope) more clear, and contains
an explanation of why the checks are needed. Also if the aggregates cannot be combined, the code will now do multiple db queries instead
of throwing an error.