Data table: Allow for multiple values in a single cell
This can only occur if not all dimensions of the dataset are specified as either row or column dimensions,
e.g. if the patient dimension is omitted but the dataset still includes multiple patients. In the JSON result
the value then becomes a list of values instead of a single basic value.