Revert accepting an OutputStream instead of ZipOutputStream
Fix TMT-210

Using FileOutputStream instead of ZipOutputStream in the
ExportJobExecutor was introduced in this commit:
https://github.com/thehyve/transmart-
core/commit/54766f9c9f9f2063c2c8472cbb97f18666e3afc3

This caused issues with closing the ZipOutputStream. The stream has to
be closed not only after the export job is done, but also when there are
some errors or job is interupted on purpose. This is problematic when
the FileOutputStream is wrapped in ZipOutputStream somewhere else than
ExportJobExecutor. Reverting the change was the easiest solution.